nexusdash/requirements.txt
Django==1.6.5
django-bootstrap3==4.8.2
django-debug-toolbar==1.2.1
python-nvd3==0.12.2
lxml==3.3.5
xmltodict==0.9.0
South==1.0
django-celery==3.1.10
pycrypto==2.6.1
xmltodict==0.9.0